update logic Localstorage, đăng nhập đăng xuất

+++ src/pages/RegisterPage.jsx
@@ -1,64 +1,67 @@
-import React, { useState } from "react";
-import Footer from "../components/layout/Footer";
-import NavBar from "../components/layout/NavBar";
-import styled from "styled-components";
-import { Link, useNavigate } from "react-router-dom";
-import { useDispatch, useSelector } from "react-redux";
-import { userAction } from "../store";
+import React, { useState } from 'react';
+import Footer from '../components/layout/Footer';
+import NavBar from '../components/layout/NavBar';
+import styled from 'styled-components';
+import { Link, useNavigate } from 'react-router-dom';
+import { useDispatch, useSelector } from 'react-redux';
+import { userAction } from '../store';
 function RegisterPage() {
   const dispatch = useDispatch();
 
   const navigate = useNavigate();
   const [inputs, setInputs] = useState({
-    fullName: "",
-    email: "",
-    password: "",
-    phone: "",
+    fullName: '',
+    email: '',
+    password: '',
+    phone: '',
+    carts: [],
   });
 
   const [errors, setErrors] = useState({
-    fullName: "",
-    email: "",
-    password: "",
-    phone: "",
+    fullName: '',
+    email: '',
+    password: '',
+    phone: '',
   });
+
+  const [existError, setExistError] = useState('');
 
   const [formIsValid, setFormIsValid] = useState(false);
 
   //validate form đăng ký
   const validateInputs = (name, value) => {
     switch (name) {
-      case "fullName":
+      case 'fullName':
         setErrors({
           ...errors,
-          fullName: !value.trim().includes(" ")
-            ? "Điền họ tên đầy đủ của bạn!"
-            : "",
+          fullName: !value.trim().includes(' ')
+            ? 'Điền họ tên đầy đủ của bạn!'
+            : '',
         });
         break;
-      case "email":
+      case 'email':
         setErrors({
           ...errors,
           email:
-            !value.includes("@") || !value.includes(".")
-              ? "Email không hợp lệ"
-              : "",
+            !value.includes('@') || !value.includes('.')
+              ? 'Email không hợp lệ'
+              : '',
         });
         break;
-      case "password":
+      case 'password':
         setErrors({
           ...errors,
-          password: value.length < 8 ? "Mật khẩu phải có ít nhất 8 ký tự" : "",
+          password: value.length < 8 ? 'Mật khẩu phải có ít nhất 8 ký tự' : '',
         });
         break;
 
-      case "phone":
+      case 'phone':
         setErrors({
           ...errors,
           phone:
-            isNaN(value) || !value[0] === "+"
-              ? "Số điện thoại chưa hợp lệ"
-              : "",
+            isNaN(value) || !value[0] === '+'
+              ? 'Số điện thoại chưa hợp lệ'
+              : '',
         });
         break;
       default:
@@ -66,6 +69,7 @@ function RegisterPage() {
     }
   };
   const changeHandler = (e) => {
+    setExistError('');
     const { name, value } = e.target;
     validateInputs(name, value);
     setInputs({
@@ -75,8 +79,8 @@ function RegisterPage() {
 
     //kiểm tra validate thành công và tất cả các trường input đã được nhập thì mới chuyển trạng thái của form sang valid
     if (
-      Object.values(errors).every((err) => err === "") &&
-      Object.values(inputs).every((inp) => inp !== "")
+      Object.values(errors).every((err) => err === '') &&
+      Object.values(inputs).every((inp) => inp !== '')
     ) {
       setFormIsValid(() => true);
     } else {
@@ -88,31 +92,48 @@ function RegisterPage() {
   const submitHandler = (e) => {
     //ngăn mặc định
     e.preventDefault();
-    //lấy thông tin người dùng
+    //lấy thông tin người dùng từ localstorage
 
-    const userArr = JSON.parse(localStorage.getItem("USERARR")) || [];
+    // const userArr = JSON.parse(localStorage.getItem("USERARR")) || [];
 
-    const newUserArr = [...userArr, inputs];
-    //lưu vào localstorage
-    localStorage.setItem("USERARR", JSON.stringify(newUserArr));
-    console.log(inputs, errors, userArr);
+    const userArr = localStorage.getItem('USERARR')
+      ? JSON.parse(localStorage.getItem('USERARR'))
+      : [];
 
-    //đặt các giá trị về ban đầu
-    setInputs({
-      fullName: "",
-      email: "",
-      password: "",
-      phone: "",
-    });
-    setErrors({
-      fullName: "",
-      email: "",
-      password: "",
-      phone: "",
-    });
-    setFormIsValid(false);
-    //chuyển hướng sang trang login
-    navigate("/login", { replace: false });
+    for (let i = 0; i < userArr.length; i++) {
+      if (userArr[i].email === inputs.email) {
+        console.log('email trùng!', i, userArr[i].email);
+        setExistError('Email đã được sử dụng!Vui lòng kiểm tra lại!');
+        setFormIsValid(false);
+        // break;
+        return;
+      }
+    }
+
+    if (!existError) {
+      const newUserArr = [...userArr, inputs];
+      //lưu vào localstorage
+      localStorage.setItem('USERARR', JSON.stringify(newUserArr));
+      console.log(inputs, errors, userArr, existError);
+
+      //đặt các giá trị về ban đầu
+      setInputs({
+        fullName: '',
+        email: '',
+        password: '',
+        phone: '',
+        carts: [],
+      });
+      setErrors({
+        fullName: '',
+        email: '',
+        password: '',
+        phone: '',
+      });
+      setFormIsValid(false);
+      //chuyển hướng sang trang login
+      navigate('/login', { replace: false });
+    }
   };
 
   return (
@@ -141,6 +162,7 @@ function RegisterPage() {
           onChange={changeHandler}
         />
         {errors.email && <p className="error">{errors.email}</p>}
+        {existError && <p className="error">{existError}</p>}
         <input
           type="password"
           name="password"
@@ -164,7 +186,7 @@ function RegisterPage() {
         </button>
       </form>
       <p className="link-to__login">
-        Login? <Link to={"/login"}>Click</Link>
+        Login? <Link to={'/login'}>Click</Link>
       </p>
       <Footer />
     </RegisterWrapper>
@@ -177,7 +199,7 @@ const RegisterWrapper = styled.div`
   max-width: 1200px;
   margin: auto;
   // background-image: url('${(props) => props.backgroundImg}');
-  background: url("${(props) => props.backgroundImg}");
+  background: url('${(props) => props.backgroundImg}');
   height: 500px; /* You must set a specified height */
   background-position: center; /* Center the image */
   background-repeat: repeat; /* Do not repeat the image */
